fix up htdigest test
git-svn-id: http://code.sixapart.com/svn/djabberd/trunk@367 43dd9337-660f-0410-bd32-e7601923a1a1

+++ t/lib/djabberd-test.pl
@@ -263,7 +263,7 @@ sub start {
 
 sub kill {
     my $self = shift;
-    CORE::kill(9, $self->{pid});
+    CORE::kill(9, $self->{pid}) if $self->{pid};
 }
 
 sub DESTROY {
